refactor(Piece): 클래스 상속 계층 세분화
