+++ src/main/java/chess/domain/BoardFactory.java
@@ -4,10 +4,10 @@ import chess.domain.piece.Bishop;
 import chess.domain.piece.BlackPawn;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
-import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
 import chess.domain.piece.WhitePawn;
+import chess.domain.piece.abstractPiece.Piece;
 import chess.domain.piece.character.Team;
 import java.util.HashMap;
 import java.util.Map;
