+++ src/main/java/chess/dto/BoardStatusDto.java
@@ -2,11 +2,12 @@ package chess.dto;
 
 import chess.domain.Position;
 import chess.domain.Status;
-import chess.domain.piece.Piece;
+import chess.domain.piece.abstractPiece.Piece;
 import chess.view.CharacterViewer;
 import java.util.Map;
 
 public class BoardStatusDto {
+    private static final String EMPTY = ".";
     private final Map<Position, Piece> board;
     private final Status status;
 
@@ -19,7 +20,7 @@ public class BoardStatusDto {
         if (board.containsKey(Position.of(row, column))) {
             return CharacterViewer.convertToString(board.get(Position.of(row, column)).character());
         }
-        return ".";
+        return EMPTY;
     }
 
     public Status status() {
